+++ HISTORY.rst
@@ -6,6 +6,7 @@ History
 v0.12.2 (XXXX-XX-XX)
 ....................
 * raise an exception if a field's name shadows an existing ``BaseModel`` attribute #242
+* add ``UrlStr`` and ``urlstr`` types #236
 
 v0.12.1 (2018-07-31)
 ....................